[#noissue] refactoring move mongo transformers to a dedicated class
